The output directory for the web app was changed and cautionary notes were added.
The 'WebApp' output directory was specified to achieve the writing of the JSON file. For some unknown reason, this file is bigger than 100 MB and cannot be uploaded to GitHub. Each new feature must be mapped in this file and manually pasted into the 300 kB JSON file in the main folder.

Caution notes were added in the interface to inform users that no data is stored, that they need to complete the process, and that working with a new table involves refreshing the browser.